refactor GpuHiveTableScanExec shims
